FOR-196 Extends editor to support JSON variations
Allows the editor to handle variations of JSON files, specifically those with a '.pcore.json' extension.

This involves updating the file selector in 'package.json' to include this new pattern and adjusting the activation condition for the relevant command to include both '.pcore' and '.json' extensions.

Additionally, this change fixes a typo in the PCoreDocument class by renaming 'extention' to 'extension' and updates the tsconfig to use "node16" module resolution.